Adding new option
Adding new option to TOC ( `addBottomPadding` ) which receives a jQuery selector.
When an element is clicked, TOC will check wether there is enough space on the document / window for the target to be positioned at the top of the screen. If there is not, it will add the padding it needs to do it.